add more integration tests
