+++ integration_tests/family/test_family_get.py
@@ -1,0 +1,95 @@
+from fastapi.testclient import TestClient
+from fastapi import status
+from sqlalchemy.orm import Session
+from integration_tests.family.family_helpers import EXPECTED_FAMILIES, setup_db
+
+
+# --- GET ALL
+
+
+def test_get_all_families(client: TestClient, test_db: Session, user_header_token):
+    setup_db(test_db)
+    response = client.get(
+        "/api/v1/families",
+        headers=user_header_token,
+    )
+    assert response.status_code == status.HTTP_200_OK
+    data = response.json()
+    assert type(data) is list
+    assert len(data) == 3
+    ids_found = set([f["import_id"] for f in data])
+    expected_ids = set(["A.0.0.1", "A.0.0.2", "A.0.0.3"])
+
+    assert ids_found.symmetric_difference(expected_ids) == set([])
+
+    sdata = sorted(data, key=lambda d: d["import_id"])
+    assert sdata[0] == EXPECTED_FAMILIES[0]
+    assert sdata[1] == EXPECTED_FAMILIES[1]
+    assert sdata[2] == EXPECTED_FAMILIES[2]
+
+
+def test_get_all_families_is_authed(client: TestClient, test_db: Session):
+    setup_db(test_db)
+    response = client.get(
+        "/api/v1/families",
+    )
+    assert response.status_code == status.HTTP_401_UNAUTHORIZED
+
+
+# --- GET
+
+
+def test_get_family(client: TestClient, test_db: Session, user_header_token):
+    setup_db(test_db)
+    response = client.get(
+        "/api/v1/families/A.0.0.1",
+        headers=user_header_token,
+    )
+    assert response.status_code == status.HTTP_200_OK
+    data = response.json()
+    assert data["import_id"] == "A.0.0.1"
+    assert data == EXPECTED_FAMILIES[0]
+
+
+def test_get_family_is_authed(client: TestClient, test_db: Session):
+    setup_db(test_db)
+    response = client.get(
+        "/api/v1/families/A.0.0.1",
+    )
+    assert response.status_code == status.HTTP_401_UNAUTHORIZED
+
+
+def test_get_family_404(client: TestClient, test_db: Session, user_header_token):
+    setup_db(test_db)
+    response = client.get(
+        "/api/v1/families/A.0.0.8",
+        headers=user_header_token,
+    )
+    assert response.status_code == 404
+    data = response.json()
+    assert data["detail"] == "Family not found: A.0.0.8"
+
+
+def test_get_family_400(client: TestClient, test_db: Session, user_header_token):
+    setup_db(test_db)
+    response = client.get(
+        "/api/v1/families/A008",
+        headers=user_header_token,
+    )
+    assert response.status_code == 400
+    data = response.json()
+    expected_msg = "The import id A008 is invalid!"
+    assert data["detail"] == expected_msg
+
+
+def test_get_family_503(
+    client: TestClient, test_db: Session, bad_family_repo, user_header_token
+):
+    setup_db(test_db)
+    response = client.get(
+        "/api/v1/families/A.0.0.8",
+        headers=user_header_token,
+    )
+    assert response.status_code == 503
+    data = response.json()
+    assert data["detail"] == "Bad Repo"
