+++ unit_tests/routers/test_family.py
@@ -3,6 +3,7 @@ Tests the routes for family management.
 
 This uses a service mock and ensures each endpoint calls into the service.
 """
+from fastapi import status
 from fastapi.testclient import TestClient
 from app.model.family import FamilyDTO
 import app.service.family as family_service
@@ -13,7 +14,7 @@ def test_get_all_families_uses_service_200(
     client: TestClient, family_service_mock, user_header_token
 ):
     response = client.get("/api/v1/families", headers=user_header_token)
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     data = response.json()
     assert type(data) is list
     assert len(data) > 0
@@ -25,7 +26,7 @@ def test_get_family_uses_service_200(
     client: TestClient, family_service_mock, user_header_token
 ):
     response = client.get("/api/v1/families/import_id", headers=user_header_token)
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     data = response.json()
     assert data["import_id"] == "import_id"
     assert family_service.get.call_count == 1
@@ -45,7 +46,7 @@ def test_search_family_uses_service_200(
     client: TestClient, family_service_mock, user_header_token
 ):
     response = client.get("/api/v1/families/?q=anything", headers=user_header_token)
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     data = response.json()
     assert type(data) is list
     assert len(data) > 0
@@ -68,7 +69,7 @@ def test_update_family_uses_service_200(
 ):
     new_data = create_family_dto("fam1").dict()
     response = client.put("/api/v1/families", json=new_data, headers=user_header_token)
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     data = response.json()
     assert data["import_id"] == "fam1"
     assert family_service.update.call_count == 1
@@ -90,7 +91,7 @@ def test_create_family_uses_service_200(
 ):
     new_data = create_family_dto("fam1").dict()
     response = client.post("/api/v1/families", json=new_data, headers=user_header_token)
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     data = response.json()
     assert data["import_id"] == "fam1"
     assert family_service.create.call_count == 1
@@ -114,7 +115,7 @@ def test_delete_family_uses_service_200(
     client: TestClient, family_service_mock, admin_user_header_token
 ):
     response = client.delete("/api/v1/families/fam1", headers=admin_user_header_token)
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     assert family_service.delete.call_count == 1
 
 
@@ -122,7 +123,7 @@ def test_delete_family_fails_if_not_admin(
     client: TestClient, family_service_mock, user_header_token
 ):
     response = client.delete("/api/v1/families/fam1", headers=user_header_token)
-    assert response.status_code == 403
+    assert response.status_code == status.HTTP_403_FORBIDDEN
     assert family_service.delete.call_count == 0
